feat: Migrate components to new codebase (#2)

+++ apps/server/src/trpc/routers/k8s/procedures/delete/index.ts
@@ -1,18 +1,19 @@
-import { ERROR_K8S_CLIENT_NOT_INITIALIZED } from "../../errors";
+import { isCoreKubernetesResource } from "@/clients/k8s";
 import { protectedProcedure } from "@/trpc/procedures/protected";
-import * as k8s from "@kubernetes/client-node";
+import { CustomObjectsApi } from "@kubernetes/client-node";
+import { k8sOperation, k8sResourceConfigSchema } from "@my-project/shared";
 import { TRPCError } from "@trpc/server";
 import { z } from "zod";
+import { ERROR_K8S_CLIENT_NOT_INITIALIZED } from "../../errors";
 
 export const k8sDeleteItemProcedure = protectedProcedure
   .input(
     z.object({
       clusterName: z.string(),
-      group: z.string(),
-      version: z.string(),
       namespace: z.string(),
-      resourcePlural: z.string(),
       name: z.string(),
+      resource: z.any(),
+      resourceConfig: k8sResourceConfigSchema,
     })
   )
   .mutation(async ({ input, ctx }) => {
@@ -22,18 +23,24 @@ export const k8sDeleteItemProcedure = protectedProcedure
       throw new TRPCError(ERROR_K8S_CLIENT_NOT_INITIALIZED);
     }
 
-    const customObjectsApi = K8sClient.KubeConfig.makeApiClient(
-      k8s.CustomObjectsApi
-    );
+    const { name, namespace, resourceConfig } = input;
 
-    const { group, version, namespace, resourcePlural, name } = input;
-    const res = await customObjectsApi.deleteNamespacedCustomObject({
-      group,
-      version,
-      plural: resourcePlural,
+    if (isCoreKubernetesResource(resourceConfig)) {
+      return K8sClient.callCoreResourceOperation(resourceConfig.kind, {
+        operation: k8sOperation.delete,
+        namespace,
+        name,
+      });
+    }
+
+    const customObjectsApi =
+      K8sClient.KubeConfig.makeApiClient(CustomObjectsApi);
+
+    return await customObjectsApi.deleteNamespacedCustomObject({
+      group: resourceConfig.group,
+      version: resourceConfig.version,
+      plural: resourceConfig.pluralName,
       namespace,
       name,
     });
-
-    return res;
   });
